move and change the test scenarios since we will never convert track 1 to module again
