+++ tools/common/packages/track1/packages_test.go
@@ -10,20 +10,13 @@ import (
 )
 
 const (
-	testRoot = "../../testpkgs"
+	testRoot = "./testpkgs"
 )
 
 var (
 	expected = map[string]string{
 		"scenrioa/foo":                    "foo",
-		"scenriob/foo":                    "foo",
-		"scenriob/foo/v2":                 "foo",
-		"scenrioc/mgmt/2019-10-11/foo":    "foo",
-		"scenriod/mgmt/2019-10-11/foo":    "foo",
-		"scenriod/mgmt/2019-10-11/foo/v2": "foo",
-		"scenrioe/mgmt/2019-10-11/foo":    "foo",
-		"scenrioe/mgmt/2019-10-11/foo/v2": "foo",
-		"scenrioe/mgmt/2019-10-11/foo/v3": "foo",
+		"scenriob/mgmt/2019-10-11/foo":    "foo",
 	}
 )
