+++ tools/common/packages/track1/testpkgs/scenriob/mgmt/2019-10-11/foo/fooapi/interfaces.go
@@ -2,7 +2,7 @@ package fooapi
 
 import (
 	"context"
-	"github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenrioa/foo"
+	"github.com/Azure/azure-sdk-for-go/tools/common/packages/track1/testpkgs/scenrioa/foo"
 )
 
 // GatewaysClientAPI ...
