--- tools/testpkgs/scenriob/foo/v2/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenarioa/foo/v2
-
-go 1.12
