--- tools/testpkgs/scenrioc/mgmt/2019-10-11/foo/model.go
@@ -1,12 +1,0 @@
-package foo
-
-import "github.com/Azure/go-autorest/autorest"
-
-// Gateway ...
-type Gateway struct {
-	autorest.Response `json:"-"`
-	// Field1 ...
-	Field1 *string
-	// Field2 ...
-	Field2 *int
-}
