--- tools/testpkgs/scenriod/mgmt/2019-10-11/foo/fooapi/interfaces.go
@@ -1,11 +1,0 @@
-package fooapi
-
-import (
-	"context"
-	"github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenrioa/foo"
-)
-
-// GatewaysClientAPI ...
-type GatewaysClientAPI interface {
-	CreateOrUpdate(ctx context.Context, resourceGroupName string, gatewayName string, parameters foo.Gateway) (result foo.Gateway, err error)
-}
