--- tools/testpkgs/scenrioe/mgmt/2019-10-11/foo/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenarioa/foo
-
-go 1.12
