--- tools/testpkgs/scenrioe/mgmt/2019-10-11/foo/v3/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/testpkgs/scenarioa/foo/v3
-
-go 1.12
